+++ src/interfaces/entity-options.interface.ts
@@ -1,3 +1,3 @@
 export interface EntityOptions {
-  options: string[]
+  options?: string[]
 }